Make NROFF version of manpage master, remove HTML
Vacation had been generating the NROFF version of the manual
page from an HTML version, but the HTML version was out of date
and the NROFF version in git was current.

Fix this up by just installing the current NROFF version and
remove the HTML version and html2man Perl script.