Moved English man page, vacation.man symlink to it
After adding the German man page it seemed like a good idea to
move the English one to the same convention (vacation-en.man)
and then make the old vacation.man a symlink to the English page.

This was a user can just change the symlink prior to installing
Vacation to get the German manual page should they so wish.