Add "make gitpackage" target for the maintainers.
To make it easier to roll a tarball for release or distribution
I've added a "gitpackage" make target which (if you have git
installed and are in a git managed directory) will roll an
appropriately named tarball in the directory above it.

It does rely on the tags being vacation-$version, so remember that!

Also fixed up the setting of the OS and ARCH variables, they
were never getting set at all from the uname commands as I
wasn't using the right syntax for GNU make!  Now works as I'm
using the $(shell cmd) notation.  Hopefully this won't break
on the *BSD's!